browser-object-dom-node: new browser object, bound to * *
This browser object invokes the hinting system to let you select any
dom node.  Useful for those "hard to reach places" when you need to
follow, copy, or focus a node that none of the other browser objects
cover.  For example, focusing html:div elements with scrollable
content, copying paragraphs, or simulating a click on fake hyperlinks
which use event listeners for their behavior.  All this and more
can be yours, just a * * push away.