move input-mode framework into buffer.js, eliminating circular dependency
This commit fixes http://bugs.conkeror.org/issue160 which was caused by
a frameset document in which a form existed in one frame, whose target
was another frame.  When the form was submitted, with focus still on the
input-box, results would be loaded in this other frame.  That load would
trigger a location-change, which would in turn trigger conkeror
erroneously switching to normal-input-mode.  The fix was to only use the
function content_buffer_update_input_mode_for_focus to determine the
proper mode, and never calling normal_input_mode directly.

added create_buffer_early_hook, fixes http://bugs.conkeror.org/issue102

under certain conditions, it is possible for a document title to be
changed before the buffer constructor has run to completion.  This causes
tab_bar_update_buffer_title to get run before tab_bar_add_buffer has been
run, so buffer.tab does not yet exist.

This commit begins to establish the idea of modular input-mode systems.
There are further changes that need to be made in order to complete that
project, and many of them have been marked with explanatory XXX and
FIXME notes.