removing M-<arrow> bindings (focus-next-link, focus-previous-link)
The original feature request for these bindings, which were meant to be
like Opera, was mistaken.  Opera uses shifted arrow keys for focusing
links.  The commands focus-next-link and focus-previous-link will still be
available for people to bind as they want in their rc, but as the impetus
for putting these commands on these keys was in error, we will keep these
keys free for the future for now.