fix issue189 (help buffer titles wrong in tab-bar)
Error was introduced in 1e1a96da42d12b107dc7c6c0041a355f9763c7a1, in which
tab-bar and new-tabs initialization was moved to create_buffer_early_hook.
The problem was that when create_buffer_early_hook runs, at the end of the
`buffer' constructor, the constructor for the subclass has not completed
running, so its fields are not filled in.  At this point in construction,
the field this.key_sequence is truly "undefined".

The solution was to put a title updater for the tab onto
create_buffer_hook, which runs when all constructors have completed.