content-handler rewrite
The main thing in this patch is a restructuring of download_helper which
makes each individual action for content of unknown mime types available
as its own function.  The user can then configure automation on a per-
mime-type basis, with the new variable, content_handlers.

mime_type_external_handlers was removed, and replaced by a rewritten api
embodied in the variable external_content_handlers.  This was done in
order to generalize the api used by that variable to make it also
useful for the new content_handlers variable.  The old api was also
overcomplex.  See http://bugs.conkeror.org/issue170

The api for external_content_handlers works like this:

  external_content_handlers.set("application/pdf", "xpdf");
  external_content_handlers.set("image/*", "feh);

The api for content_handlers works similarly:

  content_handlers.set("application/pdf", content_handler_save);