keymap_lookup: take array of keymaps
keymap_lookup now expects to be passed an array of keymaps instead of just
one keymap.  This will allow keymap inheritances to be built dynamically
at runtime as an alternative to defining them only statically with the
$parent parameter to define_keymap.  This capability is a step toward
being able to have modes which install extra keymaps without having
explicit dependencies one particular input-modes.  A cleaner implementation
of page-mode keymaps will become possible, as well as an implementation of
caret-mode (and other modes which install keymaps) that does not conflict
with page-modes.