dynamic keymaps
  This patch is a refactoring of runtime keymap handling and input-modes.
Input-modes are no longer the central primitive of the runtime keymap
system.  It is no longer the case that one input-mode is always enabled at
any time in a content buffer.  Instead, we introduce the concept of
"modalities" which both install keymaps, and sometimes enable an
input-mode.  Although input-modes do have a keymap associated with them,
it is no longer necessary to use an input-mode to install a keymap.  The
main role of input-modes is to perform side-effects such as putting an
indicator of the current mode in the minibuffer.

  Page-mode keymaps no longer depend on things like
content_buffer_normal_keymap.  Keymap inheritance is built dynamically at
runtime.  Several minor changes have been made to how keymaps are
installed by page-modes, so refer to the included page-modes to see the
new style.

  quote-mode, quote-next-mode, and caret-mode now work in any buffer type.

* Breaking Changes

  follow-current, follow-current-new-buffer, follow-current-new-window,
follow-current-new-buffer-background: commands removed, because
binding-supplied browser-objects can achieve the same effect.  For
example:

   define_key(content_buffer_anchor_keymap, "d", "follow",
              $browser_object = browser_object_focused_element);

  browser_prevent_automatic_form_focus_mode has been modularized and
renamed.  To use it, put this in your rc:

   require("block-content-focus-change.js");

  The variable browser_automatic_form_focus_window_duration has been
renamed to block_content_focus_change_duration.