youporn.js: removed
Youporn-mode doesn't work, but even if it did, it serves no purpose.
It uses Conkeror's media-scraper system to find the url of the flv
file for an embedded media player.  However, youporn.com already
provides a hyperlink to download that flv file.  Additionally, even
if you do want to use the media-scraper system instead of the hyperlink
to download the file, Conkeror's built-in media_scrape_default finds
the url just fine.  In summary and conclusion, there is no reason,
in triplicate, for this mode to exist.