hints_auto_exit_delay: change default to 0
This change means that by default, the user has to hit return to exit the
hints interaction.  This change makes Conkeror's UI more predictable for
new users, less prone to run unintended commands because of extra
keystrokes outside the timeout window, and removes assumptions about
people's typing speed, browsing style, and focus of attention when using
the hints system.

To use the hints-auto-exit feature, put a line like the following in your
rc, adjusting the number to suit your preference.  500 was the old
default.

  hints_auto_exit_delay = 500;

Additionally, may people like to also have an auto-exit in the ambiguous
case:

  hints_ambiguous_auto_exit_delay = 500;

Since this is a popular feature, we are also documenting it more
prominently with the recently added contrib/config/commmon.js as well as
on the wiki.