browser-object-frames: fix error in handler
We were using a "for each/in" loop over the return value of
document.getElementsByTagName, which was incorrect, because we
are only concerned with numerically indexed properties of that
object, not its extra properties like "length".  I was unable
to isolate the underlying cause for why this only sometimes
caused an error to be thrown, and I wonder if there is a bug
in mozilla such that the "length" property of an HTMLCollection
is not always seen by a "for each/in" loop.