Enable URL completion for both bookmarks and history (workaround).
Previously, enabling url_completion_use_bookmarks when
url_completion_use_history was enabled did not provide bookmarks in
the completion list.  The underlying issue is that QUERY_TYPE_UNIFIED
is not implemented in xulrunner.

The workaround used in this patch will lead to duplicates if the URL
appears in both bookmarks and history.