Properly merge completers with null elements
When a completer does not support an element it was possible for merge
completers to erroneously call the next completer with a negative
index.  For instance, this occurred with the wikipedia webjump, for
which the search engine does not provide descriptions for its
completions.

This latent bug was exposed by 2005f2e (Let completion functions set
the match_required state) because more completers can now be on the
completion list, even if they have no results.