index-webjump: insist on completions being available
Xpath webjumps fail when no completions are available.  Previously
conkeror would just attempt to load as a url whatever input the user
had typed.  Now an error is reported and the minibuffer is exited.

This behaviour is different from other cases where match_required is
set but the input doesn't match the completions.  In those cases "No
match" is reported, but the minibuffer is not exited.  This case is
different because the completer throws an error and so does not have
an opportunity to set the match_required state.

The test suite is also adjusted to suit.