contrib: mode-line buttons for basic browser control
Simple GUI buttons are provided for controlling conkeror.  They are
intended to be unobtrusive and to steal as little screen space as
possible.  Clicking on them executes a conkeror command.  Hovering
over them tells you the command and the corresponding keystroke.

The buttons are intended to make conkeror usable for a casual user and
also to aid the novice user while they become familiar with conkeror's
interface.

They can be enabled with:
  load_paths.unshift("chrome://conkeror-contrib/content/");
  require("mode-line-buttons.js");
  add_mode_line_buttons(standard_mode_line_buttons, true);

A chrome domain for conkeror-contrib is provided.

Includes a minor enabling change to the mode-line.