load_rc: allow chrome and file uris
load_rc() will now accept a number of ways to specify the file to
load.  In particular this allows loading from chrome uris on the
command line.  The code tries to produce an nsILocalFile from whatever
it's given so that, for instance, the directory processing works for
chrome uris too.

The subscript loader will refuse to load a non-local uri.

A chrome uri is provided for the test suite, which allows usage like:

  conkeror -q -batch -l chrome://conkeror-test/content/simple

Another usage is to load a module from the contrib area:

  load_rc("chrome://conkeror-contrib/content/mode-line-buttons.js");