Added support for generating URLs with fragment indentifiers (eg #foo).
Added a new browser object class 'fragment-link', which generates
a link with a fragment identifier pointing to an element in the
current document.  This can be any element with an ID attribute, or an
A element with a NAME attribute.

Also modified load_spec_from_element to generate a load spec with a
fragment identifier when passed an element that meets this description
(and doesn't also have an HREF attribute).