session_load: fix error exposed by 6e15c57.. fix parenthesization..
the proximate problem was that correcting (! b instanceof content_buffer)
to (! (b instanceof content_buffer)) caused session_load to bail as soon
as it reached the end of the existing buffers during the window-recycling
loop, whereas before (! b instanceof content_buffer) was always a false
condition.

the ultimate problem was that session_load was coded to skip over special
buffers, skipping the loading of the corresponding session url.  (can't
load an url into a special buffer, so don't try to load the url at all)

the fix was to kill all special buffers before the recycling loop, so we
are guaranteed that all buffers are content buffers.