session_load: fix case of appending session to existing window
the changes of yesterday to make session_load deal more robustly with
special buffers failed to deal with an important case: when recycling only
some buffers in the window, or only appending the session to the window.
that patch destroyed all existing special buffers in the window, which is
clearly not the right behavior.

this patch makes session_load only destroy special buffers which are
targetted for recycling.