conkeror.xul: New Window menu item for OS X
Xulrunner still does not have working command line remoting on OS X,
so there was no way to get a new window if no windows were already open.
Since we only need one menu item, it seemed excessive and counter to
minimalism to introduce a whole new menu plus the logic to enable the
menu for only one OS, so instead we commit the small hack of repurposing
one of the special application menu items.  We repurpose the "aboutName"
item to be a "New Window" item.