new module system
This commit provides a new module system which, while in the main being
backward compatible with the old module system, provides the capability
for modules to have their own namespace.  The centerpiece of the system is
a new function 'load', which is patterned after the function of that name
in emacs lisp, plus the added capability of working with modules.

Modules now need to have as their first expression an in_module assertion,
and as their last expression a 'provide' call.  Backward compatibility is
provided by "in_module(null);" which causes the module to be loaded into
application scope instead of its own namespace.

The call_after_load procedure now takes a feature name, not a filename.
Feature names are normally the same as the filename, but without a .js
extension.

This patch fixes a bug in the theme system whereby modules in
subdirectories could not be themed.  (The theme system depends on feature
names being independent of directory structures.)

The capability to write namespaced modules is thus provided, but no
existing modules are updated in this patch.  Those can be done
incrementally as deemed useful.