fix issue1, issue38, issue243, issue259
it was a bug in getFocusedSelCtrl introduced in a1124373 (aug 22, 2008).
it seems to affect pages with iframe elements, when the root document had
focus.  a variable used in searching for the focused docshell gets
clobbered in the search, and inadvertently used later, when it no longer
has a valid value.

  Not 100% sure about issue38, as a full stack trace was not given, but it
seems likely enough to close that bug too.