fix focus problem http://bugs.conkeror.org/issue263
This problem came about because of changes to the focus system in
xulrunner 1.9.2.  See the bug report for details.

Note, Conkeror's focus-tracking system in buffer.js and minibuffer.js is
no longer necessary as of 1.9.2.  (But we still need it to retain
compatibility with earlier versions of xulrunner.)  Since xulrunner 1.9.2,
element/frame focus is tracked internally to the xul:browser, as it should
be, so all conkeror really has to do is focus the browser.