key_binding_reader: protect current keymap stack
It is possible for the keymap stack to change while the minibuffer is open
reading a key binding for describe-key or describe-key-briefly.  This patch
makes key_binding_reader clone the keymap stack, instead of assuming it
won't change.

I was able to reproduce the problem by focusing a text box and doing
'C-h c C-x u' but I cannot guarantee that this will reliably show the
bug for everyone.  Tested in XulRunner 1.9.2.4pre.