extension.js: removed
From version 1.9.3a5pre, the extension manager has been removed from
Mozilla and replaced by an Addons manager.  The way that we used the
function 'extension_is_enabled' to conditionally load glue code for
extensions like dom-inspector and adblockplus is no longer possible,
because in the new addons manager, checking for the existence of an
extension is now an asynchronous call.  Further info at the following
links:

http://bugs.conkeror.org/issue272
http://www.oxymoronical.com/blog/tag/extension-manager
http://www.oxymoronical.com/blog/2010/03/How-were-breaking-some-extensions-in-the-near-future

This change means that extension glue code must now be loaded by users
who want it, by using a 'require' call in the rc.