session_get: use buffer.display_uri_string
The display-uri is guaranteed to be the uri that the buffer is pointed to,
whether or not the document has actually started to download.  The actual
document location could be about:blank, in the instant after the uri has
been requested, and before download progress has commenced, so session
data would get lost.

resolves issue273