buffer.override_keymaps: dealt with privately inside buffer
Instead of having get_current_keymaps handle buffers in a special way, as
opposed to the minibuffer, or other possible contexts, we now handle
override_keymaps internally to the buffer, so that the outside world need
only concern itself with buffer.keymaps.  buffer.override_keymaps is now a
function --- "override" being a verb.  When called with a keymap stack, it
uses that stack to override buffer.keymaps.  When called with no argument,
it turns off the override.  The exact implementation may be up for further
tinkering.  For the moment, the only caller is quote_mode.