use Object.watch to perform minibuffer updates
saves considerable overhead with respect to editing commands in the
minibuffer.  more editing commands are now available in the minibuffer,
like upcase-word, downcase-word, capitalize-word, and cut-to-end-of-line.

resolves issue236 (C-k is cut-to-end-of-line in minibuffer)

All editing commands with the prefix "minibuffer-" have been removed,
because standard editing commands can now be used in the minibuffer.
Instead of binding editing commands in minibuffer_base_keymap or
content_buffer_text_keymap, the new keymap text_keymap is used for both
purposes.