define_key_internal: make undefine_key work with predicate bindings
there were two bugs with undefine_key when called for a predicate binding.
first, if no binding was present, a new empty binding would be added.
second, if a binding was present, it would be replaced by 'undefined'
instead of removed from the list of predicate bindings.  this patch fixes
both of those cases.