unfocus: special case for embed-selections
Determined by experiment (with xulrunner 1.9.2.10pre) that when an
html:embed has focus, there is also an active selection starting on
the embed.  This meant that when an embed was focused, you had to
hit escape twice to unfocus it: once to clear the selection, and
once to unfocus the element.   With this patch, unfocus now handles
this case specially, and both clears the selection and blurs the
embed.