default_global_keymap, content_buffer_normal_keymap: bind C-z to recurse
Some page-modes that have a lot of fallthrough key bindings that shadow
the normal navigation bindings, and as a workaround include alternative
bindings for some of these commands, for example, 'C-c f' bound to 'follow'.
That workaround breaks modularity and abstraction, and also uses keys
which by convention should be left for user bindings.  This patch fixes
this problem.

The solution employed here is to bind C-z in our main keymaps to recurse.
This way, if the 'f' key is shadowed by a page mode, you can hit 'C-z f'
to call the 'follow' command instead.

resolves issue196

gmail_base_keymap removed