read_browser_object: throw error when there is no object to read
As of this patch, null is officially defined as a valid browser object.
All commands which call read_browser_object must supply a default via
the $browser_object keyword.  The commands 'scroll' and 'reload' have
been updated to supply a default of null.  The interactive error
"No browser object" is thrown when the interactive context's browser
object is 'undefined'.