hint_digits: new variable configures numeric base of hint numbers
Hints can now be numbered in other numeric bases than 10, and with different
characters than the usual 1-9,0.  The feature is exposed through a new
user variable, hint_digits, which can be set to a string of digits like:
"abcdef".  In this instance, a means 0, b means 1, and so on.  Since the
character '0' has important uses in Conkeror's default bindings, and would
be missed if hint_digits were in effect, it is treated as special, and
will always be translated to a zero in the current numeric base.