read_buffer_show_icons: new user variable, and accompanying code
buffer icons can now be shown in the completions area during a read_buffer
session.  the feature is disabled by default.  it does not depend on favicons,
but currently, favicon.js is the only provider for buffer icons, so users
will probably want to require('favicon'); in their config to try this
feature.