new-tabs.css: change so icons won't get vertically stretched
In order to be able to vertically center icons (rather than stretch them)
in CSS, we use -moz-box-flex rather than overflow:hidden to allow labels
to be cropped.  Using overflow:hidden was (for whatever reason) not
compatible with -moz-box-align:center.  It would have worked with the dom
attribute align="center", but that is less flexible for theme authors.
The method used in this patch also means that Mozilla will also display an
ellipsis when it crops text.