chrome/chrome.manifest: move component registrations to /chrome.manifest
Linux and Mac XULRunner have an error in their implementation of
nsILocalFile.appendRelativePath where they fail to throw an error for the
presence of '..' in a relative path.  Therefore an error in the way we do
component registration has gone unnoticed until now.  Windows XULRunner
correctly throws and error for our '../components/*' entires in
chrome/chrome.manifest, causing Conkeror to not start on Windows.  This
patch moves the component registrations from chrome/chrome.manifest to the
chrome.manifest in root.  Since XULRunner 1.9.x does not read these lines,
it will only affect XULRunner 2.x, as intended.  The status of the rest of
the manifest, which uses '..' in content, skin, and locale lines, is
uncertain at this time.

https://bugzilla.mozilla.org/show_bug.cgi?id=349090