text_entry_minibuffer_state: better scrolling of default completion
Since XULRunner 2.0, a change in nsITreeBoxObject.scrollToRow caused
Conkeror's read-buffer completions list to scroll the completions in a
less than helpful way, often leaving blank rows at the bottom of the list.
This patch changes the update_completions_display method of
text_entry_minibuffer_state to center the default completion (when there
is one), but also ensure that no space is wasted on empty rows in the
completions display.