webjump: simple POST webjumps
A string webjump can now provide post_data, with replacement of '%s'
elements with the webjump arguments.  This allows simple definition
of webjumps that perform an http POST in the following form:

define_webjump("example", "http://example.com/",
               $post_data=[['search', '%s']]);

Note that only a literal '%s' is replaced, rather than that sequence
anywhere in a string; it's not clear that the latter is useful in
post_data.