youtube_scrape_text: generate default file extension
We found that mime_service.getPrimaryExtension throws
NS_ERROR_NOT_INITIALIZED on some systems for the mime type "video/webm".
What these systems had in common was no entry for this mime type in
/etc/mime.types.  This patch makes the youtube scraper generate a default
extension.  The minor part of the mime type itself is used as the default,
which should be fine for all the mime types that the youtube scraper will
encounter.

Thanks to Ben Spencer for solving this, and many others for helping to
narrow it down.