Revert "google-maps-toggle-feature: new command"
This reverts commit 9bfba9f6d7c7ce59fa8eca5a9f4fe958d0db34cc.

Made an incorrect assumption about how to locate the nodes that need to be
clicked for the secondary map features.  It doesn't really work because
the "activityId" numbers that I thought were constant actually change from
visit to visit.  It would be possible instead to use the textual content
of the nodes, but then it would only work in English.  Therefore removing
feature.  If anybody wants this feature back, it will be necessary to dig
through all of google-maps' javascript to find a reliable way to implement
this.