generate_hints: special case for sizeless anchor
A sizeless anchor may have children at different coordinates from the
anchor itself.  Our test case was the paneltoggle2 widget on google maps
(which collapses the information panel).  We do a simple thing for these
cases, and just use the client-rect of the first child of the anchor that
has one.