Quash spurious warnings on minibuffer completion.
When using the completion system many warnings with the following text
would previously be displayed on the console.

  Console error: [JavaScript Warning: "reference to undefined property this.treeBoxObject.view" {file: "chrome://global/content/bindings/tree.xml" line: 0}]
    Category: chrome javascript

The problem appears to be due to underlying XULRunner code spuriously
requiring QueryInterface to be implemented.  This workaround provides
such an implementation.

See also: https://bugzilla.mozilla.org/show_bug.cgi?id=654998
Fixes: http://bugs.conkeror.org/issue364