Minimal fix to make del.icio.us bookmark saving work again with the "new Delicious"
Not yet fixed issues: After saving the bookmark, the "new delicious"
will go to your bookmark list instead of back to the page you saved.

I suspect that adding a "jump" parameter could help here, but I
haven't found the proper value for that yet. Will fix it as soon as I
found a solution to that.