set_protocol_handlers: fix it, resolving issue372
We were experiencing a problem where simply registering a mailto handler
would later cause following a link a pdf to segfault.  Full details at bug
report url, given below.  It turned out to be caused by a careless error
in the call form for getting the service nsIExternalProtocolService.  (I
had copy-pasted that line from an example somewhere, an example which was,
it turns out, incorrect.)

Also clear the 'possibleApplicationHandlers' list, so it doesn't grow and
grow.

http://bugs.conkeror.org/issue372