Debian package: Reorder xulrunner dependencies to make 7.0 the default
Reorder xulrunner alternative dependencies so the current xulrunner
from Debian Unstable (xulrunner-7.0) is chosen by default. If
xulrunner-8.0 is installed it still gets used by default though.

It is common for packages for Debian Unstable that the first of
several alternative dependencies should always be the one from Debian
Unstable.