application: clear the startup-cache at startup
Since XULRunner 8, mozIJSSubScriptLoader.loadSubScript uses the
startup-cache to speed up loading of javascript.  Since this is what
conkeror uses for its basic 'load' procedure, this causes a problem ---
modules and user-config may be loaded from cached copies.  Mozilla does
not perform any checks to see whether the files have changed.

In light of this, we should consider whether the whole module system
should be scrapped, but for as long as we use loadSubScript, this
workaround will be necessary.  Alternatively, if Mozilla could provide
more fine-grained control over the startup-cache, to allow invalidation of
particular entries, that may also lay the way for a permanent solution.