content-policy: do not auto-load the content-policy component
Have the content-policy module load the content-policy component.  This
should resolve a sporadically occurring bug whereby the program would fail
to start with the error, "get_caller_source_code_reference is not defined"
in recent version of XULRunner (7 and 8 for certain).  The stack trace
points back to the content-policy component's use of add_hook, and since
source-code.js is loaded after hook.js, it suggests that the problem was a
race condition due to asynchronous loading of components.

See http://bugs.conkeror.org/issue376

This patch also has the beneficial effect that the content-policy
component will only be loaded for people who use it.