Debian package: Support xulrunner 9+10 in debian/conkeror.bin, drop support for unversioned xulrunner binaries
xulrunner-10.0~a2 seems to have changed the format of the GRE version
string, now also including the alpha number (this was not true for
xulrunner-9.0~a2).  This means that the check for xulrunner's version in
/usr/bin/conkeror is broken and cannot be simply adapted to 10.0:

$ dpkg-query -W xulrunner-9.0
xulrunner-9.0   9.0~b1-1
$ xulrunner-9.0 --gre-version
9.0
$ expr `xulrunner-9.0 --gre-version | cut -d . -f 1,2` '>=' 1.9
1

$ dpkg-query -W xulrunner-10.0
xulrunner-10.0  10.0~a2+20111114042039-1
$ xulrunner-10.0 --gre-version
10.0a2
$ expr `xulrunner-10.0 --gre-version | cut -d . -f 1,2` '>=' 1.9
0

Given that there is no unversioned xulrunner in Debian (see #648972),
that upstream minimum requirement is xulrunner-1.9.1 and that Ubuntu
does not ship any /usr/bin/xulrunner* at all, I would avoid at all to
check for the version number.  And I would also avoid to repeat
'xulrunner-' for each version and simply use the version number in the
for list.

Details see http://bugs.debian.org/649199