Revert "buffer: make sure first buffer in window is type="content-primary""
This reverts commit d228f092ad5e0df1b5094d85a1ca6cb9a471a230.

John J. Foerchs have found a workaround for the problem I was trying
to fix with the commit that I am reverting here. The commit was likely
to break a scrollbar support, and John J. Foerchs workaround is
therefore a better solution.

See the thread "Content-initiated new windows don't work" on the
Conkeror mailing list to find the workaround.