session: give a session_token as $opener for session-opened buffers
This allows code run at buffer-creation time to know whether the buffer
was created by a session or by other means.  The known use case for this
is with the content-delay tip --- it allows delayed loading of buffers
that were opened by a session, but normal loading for buffers opened in
the background by other means.