style/gesso: new accessory theme
The purpose of the gesso theme is to be loaded on top of the default
theme, and set '-moz-appearance: none' on those ui elements that need it
in order to be styled in other ways.  Thus 'gesso' derives its name from
its intended use as a metaphorical "priming coat" for other themes that
want to style the chrome with css that differs from the system
colors/style.